+++ src/routes/cart/+page.svelte
@@ -140,7 +140,7 @@
   }
 
   function findProductAndBrand(
-    productId: String,
+    productId: string,
   ): [GiftCardProduct | undefined, Brand | undefined] {
     const product = $giftCardProductsStore.find((product) => product.id === productId);
     const brand = $brandsStore.find((b) => b.id === product?.brandId);
